Update gpu_sparsity kernel benchmarking script
Summary:
att

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

+++ torchao/sparsity/utils.py
@@ -30,12 +30,12 @@ def create_semi_structured_tensor(
     mask_entries = [random.choice(choices) for i in range(r * c // 2)]
 
     mask = (
-        torch.tensor(mask_entries, dtype=dtype)
+        torch.tensor(mask_entries, dtype=torch.int32)
         .reshape(r, c)
         .contiguous()
     ).cuda()
-    sparse_weight = torch.rand(r, c).to(dtype).cuda() * mask
-    return sparse_weight
+    sparse_weight = torch.rand(r, c).cuda() * mask
+    return sparse_weight.to(dtype)
 
 # Observers
 class PerChannelNormObserver(UniformQuantizationObserverBase):
